requirements.txt
streamlit
openai
google.generativeai
PyPDF2
pandas
python-docx